bookstore_producer/requirements.txt
confluent-kafka
faker
mysql-connector-python
requests
pandas
minio